Cleanup of media-loader
